pc-app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
altgraph==0.17.3
    # via pyinstaller
paho-mqtt==1.6.1
    # via -r requirements.in
pefile==2023.2.7
    # via pyinstaller
pillow==9.5.0
    # via -r requirements.in
pyinstaller==5.11.0
    # via -r requirements.in
pyinstaller-hooks-contrib==2023.3
    # via pyinstaller
pypiwin32==223
    # via win10toast
pywin32==306
    # via pypiwin32
pywin32-ctypes==0.2.0
    # via pyinstaller
win10toast==0.9
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
